move helpers in iri e2e main test
