+++ test/e2e-iri/main_test.go
@@ -7,8 +7,10 @@ import (
 	"testing"
 
 	"github.com/openshift/api/features"
+	"github.com/stretchr/testify/require"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
+	configv1 "github.com/openshift/api/config/v1"
 	"github.com/openshift/machine-config-operator/test/framework"
 )
 
@@ -41,4 +43,29 @@ func skipIRITests() (bool, error) {
 	}
 
 	return false, nil
+}
+
+func skipIfNoBaremetal(t *testing.T) {
+	infra, err := framework.NewClientSet("").Infrastructures().Get(context.Background(), "cluster", v1.GetOptions{})
+	require.NoError(t, err)
+	if infra.Status.PlatformStatus.Type != configv1.BareMetalPlatformType {
+		t.Skip("Skipping non-baremetal platforms")
+	}
+}
+
+// Currently some tests are not supported in the OpenShift CI
+// environment (due the proxy settings)
+func skipIfOpenShiftCI(t *testing.T) {
+	items := []string{
+		// Specific to OpenShift CI.
+		"OPENSHIFT_CI",
+		// Common to all CI systems.
+		"CI",
+	}
+
+	for _, item := range items {
+		if _, ok := os.LookupEnv(item); ok {
+			t.Skip("Skipping OpenShift CI environment")
+		}
+	}
 }
